avanzando en el curso

+++ src/main/java/com/svillanueva/repositories/Repository.java
@@ -10,5 +10,5 @@ public interface Repository<T> {
 
     void guardar(T t) throws SQLException;
 
-    void eliminar(Integer id) throws SQLException;
+    void eliminar(Long id) throws SQLException;
 }
